Added a "tap to dismiss" gesture recognizer.
Added a “canTopToDismiss” property that will enable / disable the
ability to tap to dismiss a displayed Snackbar. This is helpful if an
indefinitely displayed Snackbar is currently visible. The tap gesture
is automatically removed if the user taps on the action button that way
the user can’t dismiss the Snackbar while a long running action is
taking place.